fix: isolate candidate benchmark smoke
